27-- Sept 2010:
New 2D function objects: edge-on-ring.cpp, edge-on-ring2side.cpp.

Revised pretty-printing code in model_object.cpp (ModelObject::PrintModelParams)
to optionally print errors (if called from PrintResults() in print_results.cpp)
and to correct for offsets in X0 and Y0 values; this is now used in PrintResults()
as well as in SaveParameters().