29 Mar 2012: Various changes to SConstruct file to allow for universal Intel binary creation on Mac OS X (32-bit + 64-bit), and also to force static linking to the GSL library (since GCC tends to link to the shared-library version even when the static version is present). Further fleshing-out of imfit_howto.tex.
Added MacArthur+2003 approximation for calculating Sersic b_n for n < 0.36 to
func_sersic.cpp and func_gen-sersic.cpp.

Added python/ directory with preliminary Python code.

12 April 2012:
Minor change to func_edge-on-disk.cpp -- changed name of parameter from "I_0" to
"L_0", to avoid the implication that this should be similar to a face-on exponential's
"I_0" value.

Updates to imfit_howto.tex.