19--26 Feb 2013:
Removed func_expdisk3dv2.cpp/h (and references to it in SConstruct), since
func_expdisk3d.cpp/h is now identical.

Modified function_object.cpp/h interface to include
CanCalculateTotalFlux() method, which default to returning false, and
TotalFlux() method, which default to returning -1.0. This allows derived
classes to calculate the total flux for a function analytically, if that
is feasible.
ModelObject::FindTotalFluxes() now checks CanCalculateTotalFlux() for
each component; if the result is "true", then it calls the components's
TotalFlux() method  instead of following the standard approach of
integrating a large model image for that component.

Added proper ftol_rel call to nmsimplex_fit.cpp, so that ftol value
applies to *relative* fit-statistic value [previously, we only had that
as an *absolute* value, so chi^2 or Cash-stat had to change by < ftol as
an absolute value in order to halt minimization.  Result is faster N-M
fitting and slightly different fits, but only the last decimal place or
two.

Updated do_psfconvolve_tests to include convolution of an HST WFPC2 (PC) image with
a SINFONI-derived PSF; added all psfconvolve test files to respository.

Updates to SConstruct file to enable searching for required and optional libraries,
and exiting (with error messages) if they are not found.