Changed bootstrap_errors.cpp to include two wrapper functions: BootstrapErrors() and BootstrapErrorsArrayOnly(). BootstrapErrors() has the same basic effect as before, including calculation and printing of sigmas and confidence intervals. BootstrapErrorsArrayOnly() just returns the full array of best-fit parameter values from the bootstrap without calculating anything else.
Both wrapper functions call the (non-public) base function BootstrapErrorsBase, which
handles the bootstrap resampling and (optionally) writing best-fit parameters to file.
The latter is now done on a line-by-line basis: after each successful fit, the best-fit
parameters are immediately written to file. This allows saving of successful fits
up until the point that a problem in the fitting process causes an abort or other
exit.

All functions now return the number of *successful* fits (i.e., where the minimization
algorithm returned > 0 as its status); parameters are saved (to file and/or output
array) only for successful fits.

Changed --modcash to --poisson-mlr (or just --mlr); changed most instances of variable
names and using "modified Cash" or "modifiedCash" or "modCash" or similar to
"Poisson MLR" or "poissonMLR" or similar.

Officially updated to version 1.1.