Updated print_results.cpp to print "reduced Poisson-MLR" value (Poisson-MLR/n_deg_freedom), in analogy to reduced-chi^2.
Modified imfit_main.cpp to print that bootstrap output was saved to file (and filename),
if that was done.

Changes to SConstruct and make_distributions.py to enable separate compilation for
Mac OS 10.6 and 10.7. "Standard" Mac binary distribution (imfit-1.x-macintel.tar.gz)
is now 64-bit only, for use on Mac OS 10.8 and higher; alternate binary distribution
(imfit-1.x-macintel_10.6-10.7.tar.gz) is fat 32-bit/64-bit for use on 10.6 and 10.7.